Reduce frontend complexity per complexity report findings
Extract useWsConnection hook from useChatSocket (677→460 lines) to
separate WebSocket transport lifecycle from session-level message
processing. Extract useSendMessage hook from App.tsx (615→510 lines)
to colocate file uploads, skill expansion, and tentative queue logic.
Group LeftSidebar's 37 flat props into DocsPickerProps and MobileProps
sub-interfaces for better maintainability.
